tests/wpt/mozilla/tests/incremental-layout/preserved-same-formatting-context-block-layout.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div>
  <p id="p1">Example1</p>
  <p id="p2">Example2</p>
  <p id="p3">Example3</p>
  <p id="p4">Example4</p>
  <p id="p5">Example5</p>
</div>

<script>
setup({explicit_done:true});

window.onload = () => {
    test((t) => {
        ServoTestUtils.forceLayout();
        p1.innerText = "Changed";

        let result = ServoTestUtils.forceLayout();

        // text fragment + linebox + p1 + div + body + root
        assert_equals(result.rebuiltFragmentCount, 6, "rebuiltFragmentCount");
        assert_equals(result.restyleFragmentCount, 0, "restyleFragmentCount");

    }, "Updating a sibling should not cause fragments to be rebuilt");

    done();
}
</script>
